requirements.txt
streamlit
streamlit-webrtc
ultralytics
av
opencv-python-headless
tensorflow-cpu
numpy
pandas